A new implementation of shadings, step 60.
DETAILS :

Define frac31 with int32_t to minimize compiler dependence.

The new code is disabled with USE_LINEAR_COLOR_PROCS in gxshade4.h .

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@4990 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gxdevcli.h
@@ -21,6 +21,7 @@
 #  define gxdevcli_INCLUDED
 
 #include "std.h"		/* for FILE */
+#include "stdint_.h"
 #include "gscompt.h"
 #include "gsdcolor.h"
 #include "gsmatrix.h"
@@ -186,7 +187,7 @@ typedef struct gx_device_anti_alias_info_s {
     int graphics_bits;		/* ditto */
 } gx_device_anti_alias_info;
 
-typedef long frac31; /* A fraction value in [-1,1]. 
+typedef int32_t frac31; /* A fraction value in [-1,1]. 
     Represents a color (in [0,1]) 
     or a color difference (in [-1,1]) in shadings. */
